Provide the ability to synchronise members on a line of code
Without this, it can be very hard to do things like loops against
dynamic conditions, because all members need to agree whether to continue
the loop, or whether a given value is one thing or another.

By running of the supplier on the leader only, and then ensuring all
members have reached that line and receive the same value, such patterns
are possible and deterministic.